requirement.txt
glcontext==3.0.0
moderngl==5.12.0
numpy==2.3.5
pygame-ce==2.5.6
